LayoutTests/accessibility/gtk/text-for-table-expected.txt

foo	bar
This tests the ability to get element text for a specified range. No text should be obtainable for tables which should not implement the text interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.stringForRange(0, -1) is ''
PASS successfullyParsed is true

TEST COMPLETE
